requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dnspython==2.6.1
easyofd==0.3.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fonttools==4.53.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
inflect==7.3.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.3.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==2.0.0
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyMySQL==1.1.1
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.3
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.52
stack-data==0.6.3
starlette==0.37.2
tornado==6.4.1
traitlets==5.14.3
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
win32-setctime==1.1.0
xmltodict==0.13.0
